Fix bugs where printf-like format strings didn't match the associated types.
Also use (unsigned long long) rather than (unsigned long) when printing some
size_t types.


git-svn-id: http://svn.openimageio.org/oiio/trunk@1303 31904243-a364-42ac-a3db-9305396518ba

+++ src/iv/imageviewer.cpp
@@ -739,7 +739,7 @@ ImageViewer::updateStatusBar ()
         return;
     }
     std::string message;
-    message = Strutil::format ("%d/%d) : ", m_current_image+1, m_images.size());
+    message = Strutil::format ("%d/%d) : ", m_current_image+1, (int) m_images.size());
     message += cur()->shortinfo();
     statusImgInfo->setText (message.c_str());
